+++ R/plot_surv_animated.r
@@ -4,14 +4,20 @@
 #' @importFrom rlang .data
 #' @export
 plot_surv_animated <- function(time, status, variable, group=NULL, data, model,
-                               cif=FALSE, na.action=options()$na.action,
+                               cif=FALSE, conf_int=FALSE, conf_level=0.95,
+                               n_boot=300, na.action=options()$na.action,
                                horizon=NULL, fixed_t=NULL, max_t=Inf,
                                slider=TRUE,
                                size=1, color="black", linetype="solid", alpha=1,
                                xlab="Time", ylab="Survival Probability",
                                title=NULL, subtitle=NULL,
                                gg_theme=ggplot2::theme_bw(),
-                               facet_args=list(), ...) {
+                               facet_args=list(), ci_alpha=0.4,
+                               kaplan_meier=FALSE, km_size=0.5,
+                               km_linetype="solid", km_alpha=1,
+                               km_color="black", km_ci=FALSE,
+                               km_ci_type="plain", km_ci_level=0.95,
+                               km_ci_alpha=0.4, ...) {
 
   data <- use_data.frame(data)
 
@@ -44,6 +50,11 @@ plot_surv_animated <- function(time, status, variable, group=NULL, data, model,
                          times=fixed_t,
                          na.action="na.fail",
                          cif=cif,
+                         event_time=time,
+                         event_status=status,
+                         conf_int=conf_int,
+                         conf_level=conf_level,
+                         n_boot=n_boot,
                          ...)
 
   # correct label
@@ -59,10 +70,51 @@ plot_surv_animated <- function(time, status, variable, group=NULL, data, model,
     p <- ggplot2::ggplot(plotdata, ggplot2::aes(x=.data$time, y=.data$est))
   }
 
-  p <- p + ggplot2::geom_step(size=size, color=color, linetype=linetype,
+  if (conf_int & slider) {
+    stop("Showing confidence intervals is currently not supported when using",
+         " slider=TRUE.")
+  } else if (conf_int) {
+    requireNamespace("pammtools")
+
+    p <- p + pammtools::geom_stepribbon(ggplot2::aes(x=.data$time,
+                                                     y=.data$est,
+                                                     ymin=.data$ci_lower,
+                                                     ymax=.data$ci_upper),
+                                        alpha=ci_alpha, inherit.aes=FALSE,
+                                        fill=color)
+  }
+
+  p <- p + ggplot2::geom_step(linewidth=size, color=color, linetype=linetype,
                               alpha=alpha) +
     ggplot2::labs(x=xlab, y=ylab, title=title, subtitle=subtitle) +
     gg_theme
+
+  # add kaplan-meier curve if specified
+  if (kaplan_meier) {
+    km_dat <- get_kaplan_meier(time=time, status=status, group=group,
+                               data=data, conf_int=km_ci,
+                               conf_type=km_ci_type, conf_level=km_ci_level,
+                               cif=cif)
+    if (km_ci & slider) {
+      stop("Showing confidence intervals is currently not supported when using",
+           " slider=TRUE.")
+    } else if (km_ci) {
+      requireNamespace("pammtools")
+
+      p <- p + pammtools::geom_stepribbon(data=km_dat,
+                                          ggplot2::aes(x=.data$time,
+                                                       y=.data$est,
+                                                       ymin=.data$ci_lower,
+                                                       ymax=.data$ci_upper),
+                                          fill=km_color, alpha=km_ci_alpha,
+                                          inherit.aes=FALSE)
+    }
+    p <- p + ggplot2::geom_step(data=km_dat, ggplot2::aes(x=.data$time,
+                                                          y=.data$est),
+                                linewidth=km_size, color=km_color,
+                                alpha=km_alpha, linetype=km_linetype,
+                                inherit.aes=FALSE)
+  }
 
   # facet plot by factor variable
   if (!is.null(group)) {
@@ -83,8 +135,8 @@ plot_surv_animated <- function(time, status, variable, group=NULL, data, model,
       title <- variable
     }
 
-    p <- p + gganimate::transition_time(.data$cont) +
-      ggplot2::labs(title=paste0(title, ": {frame_time}"))
+    p <- p + gganimate::transition_manual(frames=.data$cont) +
+      ggplot2::labs(title=paste0(title, ": {frame}"))
   }
 
   return(p)