+++ R/plot_surv_rmst.r
@@ -83,6 +83,8 @@ plot_surv_rmst <- function(time, status, variable, group=NULL,
                          horizon=horizon,
                          times=fixed_t,
                          na.action="na.fail",
+                         event_time=time,
+                         event_status=status,
                          ...)
 
   # calculate RMST values
@@ -106,10 +108,11 @@ plot_surv_rmst <- function(time, status, variable, group=NULL,
 
   if (length(tau)==1) {
     p$mapping$colour <- NULL
-    gg_line <- ggplot2::geom_line(size=size, linetype=linetype, alpha=alpha,
-                                  color=color)
+    gg_line <- ggplot2::geom_line(linewidth=size, linetype=linetype,
+                                  alpha=alpha, color=color)
   } else {
-    gg_line <- ggplot2::geom_line(size=size, linetype=linetype, alpha=alpha)
+    gg_line <- ggplot2::geom_line(linewidth=size, linetype=linetype,
+                                  alpha=alpha)
   }
 
   p <- p + gg_line +