+++ tests/testthat/test_plot_surv_area.r
@@ -8,6 +8,11 @@ sim_dat$group <- factor(sim_dat$group)
 model <- survival::coxph(survival::Surv(time, event) ~ x3 + group,
                          data=sim_dat, x=TRUE)
 
+sim_dat$x4 <- sim_dat$x3
+sim_dat$x4[sim_dat$x4 > 30] <- sim_dat$x4[sim_dat$x4 > 30] - 100
+model2 <- survival::coxph(survival::Surv(time, event) ~ splines::bs(x4, df=3),
+                          data=sim_dat, x=TRUE)
+
 test_that("plot, defaults", {
   plt <- plot_surv_area(time="time", status="event", variable="x3",
                         data=sim_dat, model=model)
@@ -20,6 +25,13 @@ test_that("plot, with group", {
                         group="group", data=sim_dat, model=model)
   expect_s3_class(plt, "ggplot")
   vdiffr::expect_doppelganger("plot, with group", fig=plt)
+})
+
+test_that("plot, with nonmonotonic", {
+  plt <- plot_surv_area(time="time", status="event", variable="x4",
+                        monotonic=FALSE, data=sim_dat, model=model2)
+  expect_s3_class(plt, "ggplot")
+  vdiffr::expect_doppelganger("plot, with nonmonotonic", fig=plt)
 })
 
 test_that("plot, cif", {
@@ -76,6 +88,15 @@ test_that("plot, sep_lines", {
   vdiffr::expect_doppelganger("plot, sep_lines", fig=plt)
 })
 
+test_that("plot, with kaplan_meier", {
+  plt <- plot_surv_area(time="time", status="event", variable="x3",
+                        data=sim_dat, model=model, discrete=TRUE,
+                        kaplan_meier=TRUE, km_size=1, km_color="white",
+                        km_alpha=0.8, km_linetype="dashed")
+  expect_s3_class(plt, "ggplot")
+  vdiffr::expect_doppelganger("plot, with kaplan_meier", fig=plt)
+})
+
 test_that("plot, lots of stuff", {
   plt <- plot_surv_area(time="time", status="event", variable="x3",
                         data=sim_dat, model=model, start_color="grey",
@@ -86,4 +107,19 @@ test_that("plot, lots of stuff", {
                         gg_theme=ggplot2::theme_classic())
   expect_s3_class(plt, "ggplot")
   vdiffr::expect_doppelganger("plot, lots of stuff", fig=plt)
+})
+
+test_that("using group with monotonic=FALSE", {
+  expect_error(plot_surv_area(time="time", status="event", variable="x3",
+                              monotonic=FALSE, group="group",
+                              data=sim_dat, model=model),
+               "The 'group' argument cannot be used with monotonic=FALSE.")
+})
+
+test_that("using discrete with monotonic=FALSE", {
+  expect_warning(plot_surv_area(time="time", status="event", variable="x3",
+                                monotonic=FALSE, discrete=TRUE,
+                                data=sim_dat, model=model),
+               paste0("To obtain valid results when using monotonic=FALSE, the",
+                      " 'horizon' should contain at least 40 distinct values."))
 })